pyproject.toml
[tool.poetry]
name = "xtracture"
version = "0.1.0"
description = ""
authors = ["ryo.ishii <ryoishii1101@gmail.com>"]
readme = "README.md"
packages = [{include = "xtracture", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.2"
langchain = "^0.0.117"
google-cloud-vision = "^3.4.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.1.1"
pre-commit = "^3.2.0"

[tool.black]
line_length = 120

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
